Refactor setup.sh script to improve readability and add support for Kali Linux and Termux

+++ setup.sh
@@ -6,7 +6,6 @@
 # Modified on : March 04, 2024
 
 # This file is part of InstaHack.
-#
 # Defining the ANSII color code variables for colored output
 RED="\033[91m"
 GREEN="\033[92m"
@@ -16,61 +15,37 @@ RED_REV="\033[07;91m"
 YELLOW_REV="\033[07;93m"
 DEFCOL="\033[00m"
 
-# The main script starts here
-clear
+# Function to detect the operating system
+detect_os() {
+    uname
+}
 
-# Detect the user's operating system
-os=$(uname)
+# Function to check the Linux distribution
+check_distro() {
+    cat /etc/*-release | grep '^ID='
+}
 
-# Check the user's operating system and perform different actions based on the operating system
-if [ "$os" = "Linux" ]; then
-    # If the user is using Linux, check the distribution
-    distro=$(cat /etc/*-release | grep '^ID=')
+# Function to install the required utilities and modules
+install_utilities_and_modules_kali() {
+    sudo apt-get install python -y >/dev/null 2>&1
+    sudo apt-get install git -y >/dev/null 2>&1
+    sudo apt-get install wget -y >/dev/null 2>&1
+    sudo apt-get install curl -y >/dev/null 2>&1
+    pip3 install lolcat -y >/dev/null 2>&1
+    sudo apt install tor -y >/dev/null 2>&1
+}
 
-    if [[ "$distro" == *"kali"* ]]; then
-        # If the user is using Kali Linux, install the required utilities and modules
-        clear
-        echo ""
-        echo -e $GREEN "Detected Kali Linux. Installing the required utilities and modules..." $DEFCOL
-        echo -e $GREEN "Please wait This may take a moment depending on your internet speed..." $DEFCOL
-        echo -e $GREEN "You may be prompted to enter your password..." $DEFCOL
-        echo ""
-        sleep 1
-        sudo apt-get install python -y >/dev/null 2>&1
-        sudo apt-get install git -y >/dev/null 2>&1
-        sudo apt-get install wget -y >/dev/null 2>&1
-        sudo apt-get install curl -y >/dev/null 2>&1
-        pip3 install lolcat -y >/dev/null 2>&1
-        sudo apt install tor -y >/dev/null 2>&1
-        git clone https://github.com/evildevill/instahack.git
-        cd instahack
-        dir="/home/$USER/.local/bin"
-        # Check if the directory is in the PATH
-        if echo $PATH | grep -q $dir; then
-            echo "$dir is already in the PATH"
-        else
-            echo "Adding $dir to PATH"
-            echo "export PATH=$dir:\$PATH" >>~/.bashrc
-            source ~/.bashrc
-        fi
-        pip3 install -r requirements.txt
-        sudo rm -rf /etc/tor/torrc
-        sudo cp torrc /etc/tor/torrc
-        echo -e $GREEN "Setup completed successfully" $DEFCOL
-        echo -e $GREEN "Open new terminal and type $RED tor $GREEN then hit enter" $DEFCOL
-        sleep 5
-        bash instahack.sh
-    else
-        echo -e $GREEN "Sorry, this Linux distribution is not supported" $DEFCOL
+# Function to add a directory to the PATH if it's not already there
+add_to_path() {
+    dir="/home/$USER/.local/bin"
+    if ! echo $PATH | grep -q $dir; then
+        echo "export PATH=$dir:\$PATH" >>~/.zshrc
+        source ~/.zshrc
     fi
-elif [ -d "/data/data/com.termux/files/usr/" ]; then
-    # If the user is using Termux on Android, install the required utilities and modules
-    clear
-    echo ""
-    echo -e $GREEN "Detected Termux on Android. Installing the required utilities and modules..." $DEFCOL
-    echo -e $GREEN "Please wait This may take a moment..." $DEFCOL
-    echo ""
-    sleep 3
+}
+
+# Function to install the required utilities and modules for Termux
+install_utilities_and_modules_termux() {
     pkg install python -y
     pkg install python3 -y
     pkg install git -y
@@ -85,15 +60,75 @@ elif [ -d "/data/data/com.termux/files/usr/" ]; then
     pip3 install requests[socks]
     pip3 install stem
     pip3 install instagram-private-api==1.6.0
-    cd
-    git clone https://github.com/evildevill/instahack.git
-    cd instahack
-    rm -rf /data/data/com.termux/files/usr/etc/tor/torrc
-    cp torrc /data/data/com.termux/files/usr/etc/tor/torrc
-    echo -e $GREEN "Setup completed successfully" $DEFCOL
-    echo -e $GREEN "Open new terminal and type $RED tor $GREEN then hit enter" $DEFCOL
-    sleep 4
-    bash instahack.sh
+}
+
+# The main script starts here
+clear
+
+# Detect the user's operating system
+operating_system=$(detect_os)
+
+# Check the user's operating system and perform different actions based on the operating system
+if [ "$operating_system" = "Linux" ]; then
+    # If the user is using Linux, check the distribution
+    linux_distro=$(check_distro)
+
+    if [[ "$linux_distro" == *"kali"* ]]; then
+        # If the user is using Kali Linux, install the required utilities and modules
+        clear
+        echo ""
+        echo -e $GREEN "Detected Kali Linux. Installing the required utilities and modules..." $DEFCOL
+        echo -e $GREEN "Please wait This may take a moment depending on your internet speed..." $DEFCOL
+        echo -e $GREEN "You may be prompted to enter your password..." $DEFCOL
+        echo ""
+        sleep 1
+        install_utilities_and_modules_kali
+        add_to_path
+        if [ -d "instahack" ]; then
+            cd instahack
+            sleep 2
+            bash instahack.sh
+        else
+            git clone https://github.com/evildevill/instahack.git
+            cd instahack
+            sudo rm -rf /etc/tor/torrc
+            sudo cp torrc /etc/tor/torrc
+            pip3 install -r requirements.txt
+            sleep 2
+            echo -e $GREEN "Setup completed successfully" $DEFCOL
+            echo -e $GREEN "Open new terminal and type $RED tor $GREEN then hit enter" $DEFCOL
+            sleep 5
+            bash instahack.sh
+        fi
+    else
+        echo -e $GREEN "Sorry, this Linux distribution is not supported" $DEFCOL
+    fi
+elif [ -d "/data/data/com.termux/files/usr/" ]; then
+    # If the user is using Termux on Android, install the required utilities and modules
+    clear
+    echo ""
+    echo -e $GREEN "Detected Termux on Android. Installing the required utilities and modules..." $DEFCOL
+    echo -e $GREEN "Please wait This may take a moment..." $DEFCOL
+    echo ""
+    sleep 3
+    install_utilities_and_modules_termux
+    cd $HOME
+    # check if instahack folder exists
+    if [ -d "instahack" ]; then
+        cd instahack
+        sleep 2
+        bash instahack.sh
+    else
+        git clone https://github.com/evildevill/instahack
+        cd instahack
+        rm -rf /data/data/com.termux/files/usr/etc/tor/torrc
+        cp torrc /data/data/com.termux/files/usr/etc/tor/torrc
+        pip3 install -r requirements.txt
+        echo -e $GREEN "Setup completed successfully" $DEFCOL
+        echo -e $GREEN "Open new terminal and type $RED tor $GREEN then hit enter" $DEFCOL
+        sleep 4
+        bash instahack.sh
+    fi
 else
     # If the user's operating system does not match any of the above options, print a message
     echo -e $GREEN "Sorry, this operating system is not supported" $DEFCOL
